requirements.txt
Django==*.*.*
line-bot-sdk==*.*.*
gunicorn==*.*.*
django-heroku==*.*.*
psycopg2-binary==*.*.*
requests==*.*.*